Move context logic to trait
